📝 Add Group docs
