+++ packages/eds-core-react/src/components/Button/Button.stories.tsx
@@ -1,14 +1,11 @@
 import { useState, useEffect } from 'react'
-import { action } from '@storybook/addon-actions'
 import {
   Button,
   ButtonProps,
-  ButtonGroupProps,
   ToggleButtonProps,
   Icon,
   EdsProvider,
   Density,
-  Menu,
   Progress,
 } from '../..'
 import { Story, ComponentMeta } from '@storybook/react'
@@ -30,7 +27,6 @@ export default {
   title: 'Inputs/Button',
   component: Button,
   subcomponents: {
-    Group: Button.Group,
     Toggle: Button.Toggle,
   },
   args: {
@@ -325,112 +321,10 @@ export const Compact: Story<ButtonProps> = () => {
       <Button variant="ghost_icon" aria-label="menu action">
         <Icon data={menu} title="Ghost icon menu"></Icon>
       </Button>
-      <Button.Group aria-label="compact actions">
-        <Button>Contained</Button>
-        <Button variant="outlined">Outlined</Button>
-        <Button variant="outlined">Outlined</Button>
-        <Button variant="outlined">Outlined</Button>
-      </Button.Group>
     </EdsProvider>
   )
 }
 Compact.decorators = [
-  (Story) => (
-    <Stack>
-      <Story />
-    </Stack>
-  ),
-]
-
-export const GroupHorizontal: Story<ButtonGroupProps> = () => (
-  <Button.Group aria-label="primary actions">
-    <Button>Contained</Button>
-    <Button variant="outlined">Outlined</Button>
-    <Button variant="outlined">Outlined</Button>
-    <Button variant="outlined">Outlined</Button>
-  </Button.Group>
-)
-GroupHorizontal.decorators = [
-  (Story) => (
-    <Stack>
-      <Story />
-    </Stack>
-  ),
-]
-
-export const GroupVertical: Story<ButtonGroupProps> = () => (
-  <Button.Group aria-label="vertical actions" vertical>
-    <Button>Contained</Button>
-    <Button variant="outlined">Outlined</Button>
-    <Button variant="outlined">Outlined</Button>
-    <Button variant="outlined">Outlined</Button>
-  </Button.Group>
-)
-GroupVertical.decorators = [
-  (Story) => (
-    <Stack>
-      <Story />
-    </Stack>
-  ),
-]
-
-export const GroupSplit: Story<ButtonGroupProps> = () => {
-  const options = ['Create task', 'Update task', 'Delete task']
-  const [isOpen, setIsOpen] = useState<boolean>(false)
-  const [anchorEl, setAnchorEl] = useState<HTMLButtonElement>(null)
-  const [selectedIndex, setSelectedIndex] = useState(0)
-
-  const handleMenuItemClick = (event: React.MouseEvent, index: number) => {
-    action('click')(event)
-    event.stopPropagation()
-    setSelectedIndex(index)
-  }
-
-  const openMenu = () => {
-    setIsOpen(true)
-  }
-
-  const closeMenu = () => {
-    setIsOpen(false)
-  }
-
-  return (
-    <Button.Group aria-label="split buttons" style={{ gap: '1px' }}>
-      <Button>{options[selectedIndex]}</Button>
-      <Button
-        ref={setAnchorEl}
-        aria-label="select task action"
-        aria-haspopup="true"
-        aria-controls="menu-default"
-        id="anchor-split"
-        onClick={() => (isOpen ? closeMenu() : openMenu())}
-        style={{ padding: '0 4px' }}
-      >
-        <Icon name="arrow_drop_down" title="arrow_down"></Icon>
-      </Button>
-      <Menu
-        open={isOpen}
-        id="menu-split"
-        aria-labelledby="anchor-split"
-        onClose={closeMenu}
-        anchorEl={anchorEl}
-      >
-        {options.map((option, index) => (
-          <Menu.Item
-            key={option}
-            disabled={index === 2}
-            onClick={(event: React.MouseEvent) =>
-              handleMenuItemClick(event, index)
-            }
-          >
-            {option}
-          </Menu.Item>
-        ))}
-      </Menu>
-    </Button.Group>
-  )
-}
-GroupSplit.decorators = [
   (Story) => (
     <Stack>
       <Story />
