+++ packages/eds-core-react/src/components/Button/ButtonGroup/ButtonGroup.stories.tsx
@@ -1,0 +1,139 @@
+import { useState, useEffect } from 'react'
+import {
+  Button,
+  ButtonGroupProps,
+  EdsProvider,
+  Density,
+  Icon,
+  Menu,
+} from '../../..'
+import { ComponentMeta, Story } from '@storybook/react'
+import { action } from '@storybook/addon-actions'
+import { Stack } from './../../../../.storybook/components'
+import page from './ButtonGroup.docs.mdx'
+
+export default {
+  title: 'Inputs/Buttons/Group',
+  component: Button.Group,
+  parameters: {
+    docs: {
+      page,
+      source: {
+        excludeDecorators: true,
+      },
+    },
+  },
+  decorators: [
+    (Story) => (
+      <Stack>
+        <Story />
+      </Stack>
+    ),
+  ],
+} as ComponentMeta<typeof Button.Group>
+
+export const Introduction: Story<ButtonGroupProps> = (args) => {
+  return (
+    <Button.Group {...args} aria-label="primary actions">
+      <Button>Contained</Button>
+      <Button variant="outlined">Outlined</Button>
+      <Button variant="outlined">Outlined</Button>
+      <Button variant="outlined">Outlined</Button>
+    </Button.Group>
+  )
+}
+
+export const Horizontal: Story<ButtonGroupProps> = () => (
+  <Button.Group aria-label="primary actions">
+    <Button>Contained</Button>
+    <Button variant="outlined">Outlined</Button>
+    <Button variant="outlined">Outlined</Button>
+    <Button variant="outlined">Outlined</Button>
+  </Button.Group>
+)
+
+export const Vertical: Story<ButtonGroupProps> = () => (
+  <Button.Group aria-label="vertical actions" vertical>
+    <Button>Contained</Button>
+    <Button variant="outlined">Outlined</Button>
+    <Button variant="outlined">Outlined</Button>
+    <Button variant="outlined">Outlined</Button>
+  </Button.Group>
+)
+
+export const Split: Story<ButtonGroupProps> = () => {
+  const options = ['Create task', 'Update task', 'Delete task']
+  const [isOpen, setIsOpen] = useState<boolean>(false)
+  const [anchorEl, setAnchorEl] = useState<HTMLButtonElement>(null)
+  const [selectedIndex, setSelectedIndex] = useState(0)
+
+  const handleMenuItemClick = (event: React.MouseEvent, index: number) => {
+    action('click')(event)
+    event.stopPropagation()
+    setSelectedIndex(index)
+  }
+
+  const openMenu = () => {
+    setIsOpen(true)
+  }
+
+  const closeMenu = () => {
+    setIsOpen(false)
+  }
+
+  return (
+    <Button.Group aria-label="split buttons" style={{ gap: '1px' }}>
+      <Button>{options[selectedIndex]}</Button>
+      <Button
+        ref={setAnchorEl}
+        aria-label="select task action"
+        aria-haspopup="true"
+        aria-controls="menu-default"
+        id="anchor-split"
+        onClick={() => (isOpen ? closeMenu() : openMenu())}
+        style={{ padding: '0 4px' }}
+      >
+        <Icon name="arrow_drop_down" title="arrow_down"></Icon>
+      </Button>
+      <Menu
+        open={isOpen}
+        id="menu-split"
+        aria-labelledby="anchor-split"
+        onClose={closeMenu}
+        anchorEl={anchorEl}
+      >
+        {options.map((option, index) => (
+          <Menu.Item
+            key={option}
+            disabled={index === 2}
+            onClick={(event: React.MouseEvent) =>
+              handleMenuItemClick(event, index)
+            }
+          >
+            {option}
+          </Menu.Item>
+        ))}
+      </Menu>
+    </Button.Group>
+  )
+}
+
+export const Compact: Story<ButtonGroupProps> = () => {
+  const [density, setDensity] = useState<Density>('comfortable')
+
+  useEffect(() => {
+    // Simulate user change
+    setDensity('compact')
+  }, [density])
+
+  return (
+    <EdsProvider density={density}>
+      <Button.Group aria-label="compact actions">
+        <Button>Contained</Button>
+        <Button variant="outlined">Outlined</Button>
+        <Button variant="outlined">Outlined</Button>
+        <Button variant="outlined">Outlined</Button>
+      </Button.Group>
+    </EdsProvider>
+  )
+}
